#5 Remove camera code
Camera code consisted of the same RAM problem as now fixed video
loading. Unlike video loading, camera is not needed now and we do
not know how we will need to use it in the future. First we need to
find actual tiredness indicators. Full discussion can be found
in the pull request #6 (issue #3).